feat(gateway): 模型修饰符通用化 + reasoning 参数归一（吸收 new-api / OpenRouter）
两件事，都是同一个毛病的两面：把"我不懂"当成"它不存在"。

§11.25 模型修饰符通用化与变体定价
- `base@key:value` 通用语法取代四个写死的连字符后缀，规范名与书写顺序无关
  （键升序、同键后写覆盖），旧写法 `-high` 归一到 `@effort:high`，两种拼法一条账。
- 路由名与计价名分开：计价用规范变体名，路由与上游请求用基座 canonical
  ——渠道声明的是基座模型名，拿变体名去选候选会一个都选不到。
- 变体没配价回退基座价；不认识的修饰符键报模型不存在，不装懂。

§11.26 reasoning 参数归一
- 补上主流网关真正的主路：请求体参数。`reasoning_effort`（OpenAI）/
  `thinking`（Anthropic）/ 统一 `reasoning:{effort,max_tokens,enabled}`（OpenRouter 形状）
  归一成一个内部指令，再按渠道方言三向展开。
- 修一处静默失效：跨方言请求转换是严格白名单，客户端的 `reasoning_effort` 到不了
  anthropic 上游——请求 200、思考从没开过、钱照收；反向同样（预算档丢在 apply_openai）。
  `gateway_reasoning_param.rs` 先按修复前跑过一遍，五条全红。
- 三态 Unspecified/Disabled/Enabled：并成两态就关不掉模型名里带的思考。
- 参数优先于模型名后缀，但不改计费名（与 OpenRouter 一致：只有 slug 变体改计价）。
- `strip_unified` 只摘走已消化的键：没实现的 `exclude`、没看懂的 `xhigh` 原样送出去。
- 行为变更：`-thinking-N` 打到 openai 渠道此前不注入任何东西，现在按逆映射折成档位。

+++ bins/okapi/tests/gateway_routing_prefs.rs
@@ -94,7 +94,14 @@ async fn setup() -> Bed {
         let m = model.clone();
         async move {
             let (cid, _) = okapi_store::provision::create_channel(
-                &pg, &name, "openai", &base, "cred", &[m.as_str()], true, None,
+                &pg,
+                &name,
+                "openai",
+                &base,
+                "cred",
+                &[m.as_str()],
+                true,
+                None,
             )
             .await
             .unwrap();
@@ -179,7 +186,7 @@ async fn chat(bed: &Bed, provider: Option<Value>) -> (u16, Value) {
 }
 
 /// 等结算落库并返回最新一笔（结算是响应返回后的后台任务）。
-async fn last_record(pg: &PgPool, user_id: i64, n: i64) -> (i64, i64) {
+async fn last_record(pg: &PgPool, user_id: i64, n: usize) -> (i64, i64) {
     for _ in 0..80 {
         let rows = sqlx::query!(
             r#"SELECT channel_id, amount_micro FROM billing_records
@@ -189,8 +196,8 @@ async fn last_record(pg: &PgPool, user_id: i64, n: i64) -> (i64, i64) {
         .fetch_all(pg)
         .await
         .unwrap();
-        if rows.len() as i64 >= n {
-            let r = &rows[(n - 1) as usize];
+        if rows.len() >= n {
+            let r = &rows[n - 1];
             return (r.channel_id.unwrap_or(0), r.amount_micro);
         }
         tokio::time::sleep(std::time::Duration::from_millis(60)).await;
@@ -209,13 +216,18 @@ async fn zero_retention_filters_candidates() {
     assert_eq!(ch, bed.plain_channel, "缺省应投优先级更高的未声明渠道");
 
     // zdr=true：plain 被筛掉，落到声明 none 的渠道
-    assert_eq!(chat_msg(&bed, Some(json!({"zdr": true})), "z2").await.0, 200);
+    assert_eq!(
+        chat_msg(&bed, Some(json!({"zdr": true})), "z2").await.0,
+        200
+    );
     let (ch, _) = last_record(&bed.pg, bed.user_id, 2).await;
     assert_eq!(ch, bed.zero_channel, "要求零留存必须落到声明 none 的渠道");
 
     // data_collection:"deny" 是同一个诉求
     assert_eq!(
-        chat_msg(&bed, Some(json!({"data_collection": "deny"})), "z3").await.0,
+        chat_msg(&bed, Some(json!({"data_collection": "deny"})), "z3")
+            .await
+            .0,
         200
     );
     let (ch, _) = last_record(&bed.pg, bed.user_id, 3).await;
@@ -252,7 +264,10 @@ async fn max_price_rejects_before_reserving() {
     assert_eq!(status, 402, "{body}");
     assert_eq!(body["error"]["code"], "price_above_max");
     assert!(
-        body["error"]["param"].as_str().unwrap_or("").starts_with("prompt:"),
+        body["error"]["param"]
+            .as_str()
+            .unwrap_or("")
+            .starts_with("prompt:"),
         "param 要回显是哪一轴超了、实际单价多少：{body}"
     );
     assert_eq!(
@@ -271,7 +286,9 @@ async fn max_price_rejects_before_reserving() {
 
     // 上限放宽到 2.5 → 放行
     assert_eq!(
-        chat(&bed, Some(json!({"max_price": {"prompt": 2.5}}))).await.0,
+        chat(&bed, Some(json!({"max_price": {"prompt": 2.5}})))
+            .await
+            .0,
         200
     );
     // 输出侧独立判：completion_ratio 1.0 → 输出单价也是 2.0
@@ -359,7 +376,9 @@ async fn allow_fallbacks_false_stops_at_first_failure() {
 async fn provider_directive_never_reaches_upstream() {
     let bed = setup().await;
     assert_eq!(
-        chat(&bed, Some(json!({"zdr": true, "allow_fallbacks": false}))).await.0,
+        chat(&bed, Some(json!({"zdr": true, "allow_fallbacks": false})))
+            .await
+            .0,
         200
     );
     let seen = bed.seen.lock().unwrap();
